# api/openapi.yaml
openapi: 3.0.3
info:
  title: guided-ssh API
  description: |
    REST API of the guided-ssh server: certificate issuance (human + CI),
    host enrollment, grant management, and the read/audit endpoints of the
    web UI. This specification is the single source of truth; the
    Angular API client is generated from it (`make web-api`).

    Authentication: OIDC ID token as bearer token. Roles of the admin API
    (phase 8): `admin` (everything), `auditor` (read/export audit + read),
    `readonly` (read resources); roles are mapped to IdP groups and are
    inclusive of one another (admin ⊃ auditor ⊃ readonly).
  version: "1.0"
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
...
  /v1/admin/grants:
    get:
      tags: [grants]
      summary: List access rules (role readonly)
      operationId: listGrants
      responses:
        "200":
          description: All grants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Grant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [grants]
      summary: Create access rule (role admin)
      operationId: createGrant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GrantRequest" }
      responses:
        "201":
          description: Created grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Grant" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/grants/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [grants]
      summary: Read access rule (role readonly)
      operationId: getGrant
      responses:
        "200":
          description: Grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Grant" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [grants]
      summary: Update access rule (role admin)
      operationId: updateGrant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GrantRequest" }
      responses:
        "200":
          description: Updated grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Grant" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [grants]
      summary: Delete access rule (role admin)
      operationId: deleteGrant
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/admin/grants/apply:
    post:
      tags: [grants]
      summary: Declarative full reconciliation of grants (role admin)
      operationId: applyGrants
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                grants:
                  type: array
                  items: { $ref: "#/components/schemas/GrantRequest" }
      responses:
        "200":
          description: Reconciliation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplyResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/admin/ci-grants:
    get:
      tags: [ci-grants]
      summary: List CI access rules (role readonly)
      operationId: listCiGrants
      responses:
        "200":
          description: All CI grants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CiGrant" }
    post:
      tags: [ci-grants]
      summary: Create CI access rule (role admin)
      operationId: createCiGrant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CiGrantRequest" }
      responses:
        "201":
          description: Created CI grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CiGrant" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/admin/ci-grants/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [ci-grants]
      summary: Read CI access rule (role readonly)
      operationId: getCiGrant
      responses:
        "200":
          description: CI grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CiGrant" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [ci-grants]
      summary: Update CI access rule (role admin)
      operationId: updateCiGrant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CiGrantRequest" }
      responses:
        "200":
          description: Updated CI grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CiGrant" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [ci-grants]
      summary: Delete CI access rule (role admin)
      operationId: deleteCiGrant
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/admin/ci-grants/apply:
    post:
      tags: [ci-grants]
      summary: Declarative full reconciliation of CI grants (role admin)
      operationId: applyCiGrants
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ci_grants:
                  type: array
                  items: { $ref: "#/components/schemas/CiGrantRequest" }
      responses:
        "200":
          description: Reconciliation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplyResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OIDC-ID-Token
  responses:
    BadRequest:
      description: Invalid request
      content:
        text/plain: { schema: { type: string } }
    Unauthorized:
      description: Token missing or invalid
      content:
        text/plain: { schema: { type: string } }
    Forbidden:
      description: Insufficient role/permission
      content:
        text/plain: { schema: { type: string } }
    NotFound:
      description: Not found
      content:
        text/plain: { schema: { type: string } }
...
  schemas:
...
    Grant:
      type: object
      required: [id, group, issuer, tag_selector, principals, sudo, max_validity_seconds, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        group: { type: string, description: Group name in the IdP }
        issuer: { type: string }
        tag_selector:
          type: object
          additionalProperties: { type: string }
        principals: { type: array, items: { type: string } }
        sudo: { type: boolean }
        max_validity_seconds: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    GrantRequest:
      type: object
      required: [principals, max_validity_seconds]
      properties:
        group: { type: string, description: Required on POST }
        issuer: { type: string, description: empty ⇒ issuer of the admin token }
        tag_selector:
          type: object
          additionalProperties: { type: string }
        principals: { type: array, items: { type: string } }
        sudo: { type: boolean }
        max_validity_seconds: { type: integer, format: int64 }
    ApplyResult:
      type: object
      required: [created, updated, deleted, unchanged]
      properties:
        created: { type: integer }
        updated: { type: integer }
        deleted: { type: integer }
        unchanged: { type: integer }
    CiGrant:
      type: object
      required: [id, project, protected_only, tag_selector, principals, max_validity_seconds, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        project: { type: string, description: GitLab project or namespace path }
        ref_pattern: { type: string }
        protected_only: { type: boolean }
        environment_pattern: { type: string }
        tag_selector:
          type: object
          additionalProperties: { type: string }
        principals: { type: array, items: { type: string } }
        max_validity_seconds: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CiGrantRequest:
      type: object
      required: [principals, max_validity_seconds]
      properties:
        project: { type: string, description: Required on POST }
        ref_pattern: { type: string }
        protected_only: { type: boolean }
        environment_pattern: { type: string }
        tag_selector:
          type: object
          additionalProperties: { type: string }
        principals: { type: array, items: { type: string } }
        max_validity_seconds: { type: integer, format: int64 }
...